+++ deskmate-server/src/main/java/org/pwr/deskmateserver/controller/BuildingController.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import org.pwr.deskmateserver.dto.building.CreateBuildingDTO;
 import org.pwr.deskmateserver.dto.building.CreateFloorDTO;
 import org.pwr.deskmateserver.exceptions.NotFoundException;
-import org.pwr.deskmateserver.service.BuildingServiceImpl;
+import org.pwr.deskmateserver.service.BuildingService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @AllArgsConstructor
 public class BuildingController {
-    private BuildingServiceImpl buildingService;
+    private BuildingService buildingService;
 
     @GetMapping("/buildings")
     public ResponseEntity<Object> getBuildings() {
