+++ deskmate-server/src/main/java/org/pwr/deskmateserver/controller/SeatController.java
@@ -7,8 +7,6 @@ import org.pwr.deskmateserver.service.SeatService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 @RequestMapping("/api/v1/")
